split up summary section
